+++ src/video_display/sage.cpp
@@ -330,7 +330,7 @@ static struct video_frame *display_sage_getf(void *state)
         return s->frame;
 }
 
-static int display_sage_putf(void *state, struct video_frame *frame, int nonblock)
+static int display_sage_putf(void *state, struct video_frame *frame, long long nonblock)
 {
         int tmp;
         struct state_sage *s = (struct state_sage *)state;
