+++ src/video_display/unix_sock.cpp
@@ -244,7 +244,7 @@ static struct video_frame *display_unix_sock_getf(void *state)
         return vf_alloc_desc_data(s->desc);
 }
 
-static int display_unix_sock_putf(void *state, struct video_frame *frame, int flags)
+static int display_unix_sock_putf(void *state, struct video_frame *frame, long long flags)
 {
         auto s = static_cast<state_unix_sock *>(state);
         auto f = unique_frame(frame);
